Finalised roles class and instantiated within main plugin class
Added in the admin menu items for user roles. This highlighted an issue with the
defined roles. The roles were too narrow to be useful. FreeSpirit ESU needs to be
able to display Leaders, Explorers, Parents and Ex Members, however these were
further broken down for easier display on the frontend. This made it more difficult
to display on the backend so for the fine grained detail, it would be better to
use custom user fields I think, therefore the roles and capabilities were also
simplified
The initialisation of the FSESU_Roles class was added to a new function in the
base class (set_features) and the custom post type base class was also required
at this point
LICENSE.txt added
README.md added