Added a popup to display event details to the programme shortcode
Made some adjustments to the render_programme_table method to include a link
with each event listed.
Added programme.js to call the event details page in a popup window.

Note: for this to work, the main FSESU theme had to have a new Event Details
template added and the site needed a new page added at /event-details/ that uses
the Event Details template